+++ packages/dependy_flutter/lib/dependy_flutter.dart
@@ -1,3 +1,4 @@
 library dependy_flutter;
 
 export 'src/impl.dart';
+export 'src/watch.dart';